adding RestaurantAuthorizationService

+++ src/Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -8,7 +8,8 @@ using Restaurants.Domain.Entities.Restaurants;
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 public class DeleteRestaurantCommandHandler(
 	IRestaurantsDbContext dbContext,
-	ILogger<DeleteRestaurantCommandHandler> logger)
+	ILogger<DeleteRestaurantCommandHandler> logger,
+	IRestaurantAuthorizationService restaurantAuthorization)
 : IRequestHandler<DeleteRestaurantCommand, Result<Success>>
 {
 	public async ValueTask<Result<Success>> Handle(DeleteRestaurantCommand request, CancellationToken ct)
@@ -22,6 +23,15 @@ public class DeleteRestaurantCommandHandler(
 		{
 			logger.LogWarning("Restaurant with id '{id} not found'", request.Id);
 			return RestaurantErrors.RestaurantNotFound(request.Id);
+		}
+
+		bool isAuthorize = restaurantAuthorization
+				.Authorize(restaurnat, Domain.Constans.ResourceOperation.Delete);
+
+		if (!isAuthorize)
+		{
+			logger.LogWarning("Unauthorized request to delete restaurant with id '{Id}'", request.Id);
+			return Error.Forbidden(description: $"You are not the owner of the restaurant with id '{request.Id}'.");
 		}
 
 		dbContext.Restaurants.Remove(restaurnat);
